Develop

+++ src/main/java/com/github/CodeNekomancer/OADA_Backend/persistence/service/UniverseService.java
@@ -1,16 +1,21 @@
 package com.github.CodeNekomancer.OADA_Backend.persistence.service;
 
+import com.github.CodeNekomancer.OADA_Backend.configurations.ExceptionManager.NotFound.ADAccNotFoundException;
+import com.github.CodeNekomancer.OADA_Backend.configurations.ExceptionManager.NotFound.UniverseNotFoundException;
+import com.github.CodeNekomancer.OADA_Backend.configurations.XMLmanager.XmlUtil;
+import com.github.CodeNekomancer.OADA_Backend.model.UAcc.DTOs.UAccOutputDTO;
+import com.github.CodeNekomancer.OADA_Backend.model.UAcc.UAcc;
+import com.github.CodeNekomancer.OADA_Backend.model.Universe.DTOs.UniverseOutputDTO;
 import com.github.CodeNekomancer.OADA_Backend.model.Universe.Universe;
-import com.github.CodeNekomancer.OADA_Backend.model.Universe.UniverseInputDTO;
-import com.github.CodeNekomancer.OADA_Backend.model.Universe.UniverseInputDTOConverter;
+import com.github.CodeNekomancer.OADA_Backend.model.Universe.DTOs.UniverseInputDTO;
+import com.github.CodeNekomancer.OADA_Backend.model.Universe.DTOs.UniverseInputDTOConverter;
 import com.github.CodeNekomancer.OADA_Backend.persistence.repository.UniverseRepository;
-import lombok.AllArgsConstructor;
+import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.Pageable;
 import org.springframework.stereotype.Service;
 import org.w3c.dom.Document;
 import org.w3c.dom.Node;
-import org.w3c.dom.NodeList;
 import org.xml.sax.SAXException;
 
 import javax.xml.parsers.DocumentBuilder;
@@ -18,25 +23,31 @@ import javax.xml.parsers.DocumentBuilderFactory;
 import javax.xml.parsers.ParserConfigurationException;
 import java.io.IOException;
 import java.net.URL;
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
 import java.util.stream.Collectors;
 
 @Service
-@AllArgsConstructor
 public class UniverseService extends BaseService<Universe, Long, UniverseRepository> {
-    private final UniverseInputDTOConverter UIDTOC;
+    @Autowired
+    private UniverseInputDTOConverter UIDTOC;
+    @Autowired
+    private ADAccService adAccService;
 
-    public boolean genUniverse(Universe uni) {
+    public boolean genUniverse(String serverId) {
+        Universe uni = new Universe();
+        uni.setServerId(serverId);
         this.repo.save(uni);
-        modUniverseSrvc(uni);
+        modUniverseSrvc(uni.getUniverse_id());
         return this.repo.findById(uni.getUniverse_id()).isPresent();
     }
 
-    public boolean modUniverseSrvc(Universe uni) {
-        Universe u = new Universe();
+    public boolean modUniverseSrvc(Long id) {
+        if (this.repo.findById(id).isEmpty()) throw new UniverseNotFoundException();
+        Universe uni = this.repo.findById(id).get();
 
         String url = "";
-        List<String> urlList = new ArrayList<String>() {
+        List<String> urlList = new ArrayList<>() {
             {
                 add("https://s");
                 add("-");
@@ -44,14 +55,11 @@ public class UniverseService extends BaseService<Universe, Long, UniverseReposit
             }
         };
 
-        if (this.repo.findById(uni.getUniverse_id()).isPresent())
-            u = this.repo.findById(uni.getUniverse_id()).get();
-
-        if (!u.getServerId().isEmpty())
+        if (!uni.getServerId().isEmpty())
             url = urlList.get(0) +
-                    u.getServerId().substring(2) +
+                    uni.getServerId().substring(2) +
                     urlList.get(1) +
-                    u.getServerId().substring(0, 2) +
+                    uni.getServerId().substring(0, 2) +
                     urlList.get(2);
 
         DocumentBuilderFactory dbf = DocumentBuilderFactory.newInstance();
@@ -76,44 +84,31 @@ public class UniverseService extends BaseService<Universe, Long, UniverseReposit
             e.printStackTrace();
         }
 
-
         return true;
     }
 
     public boolean delUnvierseSrvc(Long id) {
+        if (this.repo.findById(id).isEmpty()) throw new UniverseNotFoundException();
         this.repo.deleteById(id);
         return !this.repo.existsById(id);
     }
 
-    public static final class XmlUtil {
-        private XmlUtil() {
-        }
-
-        public static List<Node> asList(NodeList n) {
-            return n.getLength() == 0 ?
-                    Collections.emptyList() : new NodeListWrapper(n);
-        }
-
-        static final class NodeListWrapper extends AbstractList<Node>
-                implements RandomAccess {
-            private final NodeList list;
-
-            NodeListWrapper(NodeList l) {
-                list = l;
-            }
-
-            public Node get(int index) {
-                return list.item(index);
-            }
-
-            public int size() {
-                return list.getLength();
-            }
-        }
-    }
-
-    public Page<?> getUniverseSrvc(Pageable pageable) {
+    public Page<?> getUniversePagSrvc(Pageable pageable) {
         return this.repo.findAll(pageable);
     }
 
+    public Universe getUniverseSngSrvc(Long id) {
+        if (this.repo.findById(id).isEmpty()) throw new UniverseNotFoundException();
+        return this.repo.findById(id).get();
+    }
+
+    public List<UniverseOutputDTO> getUniverseOwnSrvc(String authName) {
+        if (adAccService.findByUserName(authName).isEmpty()) throw new ADAccNotFoundException();
+
+        return  adAccService.findByUserName(authName)
+                .get()
+                .getUniverseAccounts().stream().map(UAccOutputDTO::new).collect(Collectors.toList())
+                .stream().map(UAccOutputDTO::getItsUniverse).collect(Collectors.toList())
+                .stream().filter(l -> this.repo.findById(l).isPresent()).map(l -> new UniverseOutputDTO(this.repo.getOne(l))).collect(Collectors.toList());
+    }
 }
